- Bug fix: make sure that lowest 28 GEOS-5 levels are properly remapped onto lowest 11 GEOS-4 levels. - Force YMDh integers to be real*8 for PGI compiler compatibility - Keep low hierarchy emissions in memory if high hierarchy emissions are only used within set time ranges. - Update of MEGAN e-folding times. - Allow reading nested settings/extension switches (for nested configuraton files)

+++ NcdfUtil/ncdf_mod.F90
@@ -1049,7 +1049,7 @@ CONTAINS
 !
 ! !INPUT/OUTPUT PARAMETERS:
 !
-    INTEGER,          POINTER                 :: all_YYYYMMDDhh(:)
+    INTEGER(8),       POINTER                 :: all_YYYYMMDDhh(:)
     CHARACTER(LEN=*), INTENT(  OUT), OPTIONAL :: timeUnit
     INTEGER,          INTENT(  OUT), OPTIONAL :: refYear 
 !
